refactor(provider-manager): order funcs, update names for consistency, add code docs
